+++ hc/api/tests/test_notify.py
@@ -406,6 +406,21 @@ class NotifyTestCase(BaseTestCase):
         self.assertIn("DOWN", payload["title"])
 
     @patch("hc.api.transports.requests.request")
+    def test_pushover_up_priority(self, mock_post):
+        self._setup_data("po", "123|0|2", status="up")
+        mock_post.return_value.status_code = 200
+
+        self.channel.notify(self.check)
+        assert Notification.objects.count() == 1
+
+        args, kwargs = mock_post.call_args
+        payload = kwargs["data"]
+        self.assertIn("UP", payload["title"])
+        self.assertEqual(payload["priority"], 2)
+        self.assertIn("retry", payload)
+        self.assertIn("expire", payload)
+
+    @patch("hc.api.transports.requests.request")
     def test_victorops(self, mock_post):
         self._setup_data("victorops", "123")
         mock_post.return_value.status_code = 200